requirements.txt
asgiref==3.7.2
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.2.8
django-autoslug==1.9.9
django-cleanup==8.1.0
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
packaging==23.1
pillow==10.2.0
pip-tools==7.3.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
